Move session to login and partially implement renewing login intervals
